requirements.txt
# =================================
# PocketBell Simulator Ver 1.1.1 by JA1XPM 2026/05/07
# =================================

numpy
sounddevice
websockets
pyserial
